# optimize slackutilsx.EscapeMessage function