This change is it performance related....but quite small TBH. Just found if while checking some code. 
The underlaying string.Replacer is optimized to buildup the lookup tables only once (but only if really needed!). 

```
➜  slackutilsx $ go test -bench=. --benchmem               
goos: linux
goarch: amd64
pkg: github.com/slack-go/slack/slackutilsx
cpu: Intel(R) Core(TM) i7-1065G7 CPU @ 1.30GHz
BenchmarkEscapeMessage-8      	  991929	      1230 ns/op	    6792 B/op	       9 allocs/op
BenchmarkEscapeMessageNew-8   	22310755	        58.94 ns/op	      32 B/op	       2 allocs/op
```
-> ~20 times faster, ~200 times less allocated bytes