@zchee 2022-03-11
@brainexe Thanks contribute and your change very optimized text replacer :D

Actually, I had to compare using `benchstat`, below is the result:
```
name              old time/op    new time/op    delta
EscapeMessage-20    1.54µs ± 1%    0.05µs ± 4%  -96.70%  (p=0.008 n=5+5)

name              old alloc/op   new alloc/op   delta
EscapeMessage-20    6.79kB ± 0%    0.03kB ± 0%  -99.53%  (p=0.008 n=5+5)

name              old allocs/op  new allocs/op  delta
EscapeMessage-20      9.00 ± 0%      2.00 ± 0%  -77.78%  (p=0.008 n=5+5)
```

That's would be great. I'll merge soon.